# Back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "name": "Soporte",
            "email": "soporte@terciariourquiza.edu.ar"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/horarios/disponibilidad": {
            "get": {
                "tags": [
                    "Disponibilidad"
                ],
                "summary": "Obtener todas las disponibilidades",
                "operationId": "4591fb6007f80bcc6b75b41b6945a26e",
                "responses": {
                    "200": {
                        "description": "Devuelve todas las disponibilidades"
                    },
                    "404": {
                        "description": "No se encontraron disponibilidades"
                    }
                }
            }
        },
        "/api/horarios/disponibilidad/{id}": {
            "get": {
                "tags": [
                    "Disponibilidad"
                ],
                "summary": "Obtener disponibilidad por id",
                "operationId": "d44f1bc1ba73b7f21de131ae05b30200",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilidad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve la disponibilidad"
                    },
                    "404": {
                        "description": "No se encontró la disponibilidad"
                    }
                }
            }
        },
        "/api/horarios/disponibilidad/store": {
            "post": {
                "tags": [
                    "Disponibilidad"
                ],
                "summary": "Guardar disponibilidad",
                "operationId": "d8503ca654802e7592ed29e225a69d7a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DisponibilidadDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disponibilidad guardada correctamente"
                    },
                    "500": {
                        "description": "Error al guardar la disponibilidad"
                    }
                }
            }
        },
        "/api/horarios/disponibilidad/update/{id}": {
            "put": {
                "tags": [
                    "Disponibilidad"
                ],
                "summary": "Actualizar disponibilidad",
                "operationId": "c797737ceaae1e746b597b322ce5725a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilidad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DisponibilidadDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disponibilidad actualizada correctamente"
                    },
                    "500": {
                        "description": "Error al actualizar la disponibilidad"
                    }
                }
            }
        },
        "/api/horarios/disponibilidad/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Disponibilidad"
                ],
                "summary": "Eliminar disponibilidad por id",
                "operationId": "e94afb685ab94bfb2c6822d2f48e5831",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilidad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilidad eliminada correctamente"
                    },
                    "500": {
                        "description": "Error al eliminar la disponibilidad"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DisponibilidadDTO": {
                "title": "DisponibilidadDTO",
                "description": "DTO para una disponibilidad",
                "properties": {
                    "id_uc": {
                        "description": "Id de la unidad curricular",
                        "type": "integer"
                    },
                    "id_docente": {
                        "description": "Id del docente",
                        "type": "integer"
                    },
                    "id_h_p_d": {
                        "description": "Id de la hora por dia",
                        "type": "integer"
                    },
                    "id_aula": {
                        "description": "Id del aula",
                        "type": "integer"
                    },
                    "id_grado": {
                        "description": "Id del grado",
                        "type": "integer"
                    },
                    "dia": {
                        "description": "Dia de la semana",
                        "type": "string"
                    },
                    "modulo_inicio": {
                        "description": "Modulo de inicio en formato HH:mm:ss",
                        "type": "string",
                        "format": "time"
                    },
                    "modulo_fin": {
                        "description": "Modulo de fin en formato HH:mm:ss",
                        "type": "string",
                        "format": "time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}